Add /api/v2/config endpoint with returns a list of configs
